clean: combining the example into one file for generated and runtime to make it cleaner
